fix: projectile spawns in slingshot pouch
The project's initial position is now the location of the slingshot's pounch, instead of always spawning in the middle of the slingshot's prongs. In addition, the velocity is now calculated based on the pouch's position, not the mouse cursor's position, to fix an issue where the projectile's trajectory is incorrectly calculated if the mouse is moved beyond the maximum retraction extent of the slingshot's strings.